pyproject.toml
[project]
name = "siru"
description = "Automated tests on hardware for embedded systems"
readme = "README.md"
license = "MIT"
keywords = ["embedded systems", "automated testing"]
authors = [
  { name = "Esteban Volentini", email = "evolentini@herrera.unt.edu.ar" },
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Framework :: Robot Framework",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: C",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX",
  "Operating System :: MacOS :: MacOS X",
  "Operating System :: Unix",
  "Topic :: Software Development :: Testing :: Acceptance",
]
dependencies = [
  "pyserial~=3.5",
  "pyyaml~=6.0",
  "paho-mqtt~=1.6.1"
]
dynamic = ["version"]
requires-python = ">=3.7"

[project.urls]
Documentation = "https://github.com/labmicro/siru#readme"
Issues = "https://github.com/labmicro/siru/issues"
Source = "https://github.com/labmicro/siru"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.version]
path = "src/siru/__about__.py"

[tool.hatch.envs.default]
dependencies = ["pytest", "pytest-mock", "pytest-cov"]

[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=src/siru --cov=tests {args}"
html-cov = "pytest --cov-report=html --cov-config=pyproject.toml --cov=src/siru {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.coverage.run]
branch = true
omit = ["src/siru/__about__.py"]
parallel = true

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.black]
extend-exclude = '''
'''
include = '\.pyi?$'
line-length = 88

[tool.pylint.format]
max-line-length = "88"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["src", "."]
